requirements.txt
asgiref==3.3.4
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
crispy-bootstrap5==0.6
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.3
django-allauth==0.47.0
django-bootstrap-v5==1.0.7
django-crispy-forms==1.13.0
django-decorator-include==3.0
django-environ==0.8.1
django-extra-views==0.13.0
django-filter==21.1
django-modeltranslation==0.17.3
django-translation-flags==1.0.6
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
pathspec==0.9.0
Pillow==8.2.0
platformdirs==2.4.0
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==2.3.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
reportlab==3.5.67
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.1
urllib3==1.26.5
whitenoise==5.3.0
